+++ src/main/java/edu/ie3/datamodel/io/factory/input/graphics/NodeGraphicInputFactory.java
@@ -20,16 +20,8 @@ import org.locationtech.jts.geom.Point;
 public final class NodeGraphicInputFactory
     extends GraphicInputFactory<NodeGraphicInput, NodeGraphicInputEntityData> {
 
-  private static final String POINT = "point";
-  private static final String NODE = "node";
-
   public NodeGraphicInputFactory() {
     super(NodeGraphicInput.class);
-  }
-
-  @Override
-  protected String[] getAdditionalFields() {
-    return new String[] {POINT, NODE};
   }
 
   @Override