+++ src/main/java/edu/ie3/datamodel/io/factory/timeseries/TimeBasedSimpleValueFactory.java
@@ -17,17 +17,6 @@ import java.util.*;
 
 public class TimeBasedSimpleValueFactory<V extends Value>
     extends TimeBasedValueFactory<SimpleTimeBasedValueData<V>, V> {
-  private static final String TIME = "time";
-  /* Energy price */
-  private static final String PRICE = "price";
-  /* Energy / Power */
-  private static final String ACTIVE_POWER = "p";
-  private static final String REACTIVE_POWER = "q";
-  private static final String HEAT_DEMAND = "heatDemand";
-
-  /* voltage */
-  private static final String VMAG = "vMag";
-  private static final String VANG = "VAng";
 
   public TimeBasedSimpleValueFactory(Class<? extends V> valueClasses) {
     super(valueClasses);
@@ -73,39 +62,13 @@ public class TimeBasedSimpleValueFactory<V extends Value>
       value =
           (V)
               new VoltageValue(
-                  data.getQuantity(VMAG, VOLTAGE_MAGNITUDE),
-                  data.getQuantityOptional(VANG, VOLTAGE_ANGLE));
+                  data.getQuantity(V_MAG, VOLTAGE_MAGNITUDE),
+                  data.getQuantityOptional(V_ANG, VOLTAGE_ANGLE));
     } else {
       throw new FactoryException(
           "The given factory cannot handle target class '" + data.getTargetClass() + "'.");
     }
 
     return new TimeBasedValue<>(time, value);
-  }
-
-  @Override
-  protected List<Set<String>> getFields(Class<?> entityClass) {
-    Set<String> minConstructorParams = newSet(TIME);
-
-    if (EnergyPriceValue.class.isAssignableFrom(entityClass)) {
-      minConstructorParams.add(PRICE);
-    } else if (HeatAndSValue.class.isAssignableFrom(entityClass)) {
-      minConstructorParams.addAll(Arrays.asList(ACTIVE_POWER, REACTIVE_POWER, HEAT_DEMAND));
-    } else if (HeatAndPValue.class.isAssignableFrom(entityClass)) {
-      minConstructorParams.addAll(Arrays.asList(ACTIVE_POWER, HEAT_DEMAND));
-    } else if (HeatDemandValue.class.isAssignableFrom(entityClass)) {
-      minConstructorParams.add(HEAT_DEMAND);
-    } else if (SValue.class.isAssignableFrom(entityClass)) {
-      minConstructorParams.addAll(Arrays.asList(ACTIVE_POWER, REACTIVE_POWER));
-    } else if (PValue.class.isAssignableFrom(entityClass)) {
-      minConstructorParams.add(ACTIVE_POWER);
-    } else if (VoltageValue.class.isAssignableFrom(entityClass)) {
-      minConstructorParams.addAll(List.of(VMAG, VANG));
-    } else {
-      throw new FactoryException(
-          "The given factory cannot handle target class '" + entityClass + "'.");
-    }
-
-    return Collections.singletonList(minConstructorParams);
   }
 }