+++ src/test/groovy/edu/ie3/datamodel/io/factory/input/AssetInputEntityFactoryTest.groovy
@@ -5,8 +5,14 @@
  */
 package edu.ie3.datamodel.io.factory.input
 
+import static edu.ie3.datamodel.io.naming.FieldNamingStrategy.*
+import static edu.ie3.datamodel.utils.CollectionUtils.newSet
+
 import edu.ie3.datamodel.exceptions.FactoryException
 import edu.ie3.datamodel.exceptions.NotImplementedException
+import edu.ie3.datamodel.io.naming.FieldNamingStrategy
+import edu.ie3.datamodel.io.naming.ModelFields
+import edu.ie3.datamodel.io.source.DataSource
 import edu.ie3.datamodel.models.OperationTime
 import edu.ie3.datamodel.models.input.AssetInput
 import edu.ie3.datamodel.models.input.OperatorInput
@@ -47,10 +53,10 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert operationTime == OperationTime.notLimited()
-      assert operator == operatorInput
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      operationTime == OperationTime.notLimited()
+      operator == operatorInput
+      id == parameter["id"]
     }
   }
 
@@ -73,10 +79,10 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert operationTime == OperationTime.notLimited()
-      assert operator == operatorInput
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      operationTime == OperationTime.notLimited()
+      operator == operatorInput
+      id == parameter["id"]
     }
 
     where:
@@ -106,12 +112,12 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert operationTime.startDate.present
-      assert operationTime.startDate.get() == ZonedDateTime.parse(parameter["operatesfrom"])
-      assert !operationTime.endDate.present
-      assert operator == operatorInput
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      operationTime.startDate.present
+      operationTime.startDate.get() == ZonedDateTime.parse(parameter["operatesfrom"])
+      !operationTime.endDate.present
+      operator == operatorInput
+      id == parameter["id"]
     }
   }
 
@@ -133,12 +139,12 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert !operationTime.startDate.present
-      assert operationTime.endDate.present
-      assert operationTime.endDate.get() == ZonedDateTime.parse(parameter["operatesuntil"])
-      assert operator == operatorInput
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      !operationTime.startDate.present
+      operationTime.endDate.present
+      operationTime.endDate.get() == ZonedDateTime.parse(parameter["operatesuntil"])
+      operator == operatorInput
+      id == parameter["id"]
     }
   }
 
@@ -161,13 +167,13 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert operationTime.startDate.present
-      assert operationTime.startDate.get() == ZonedDateTime.parse(parameter["operatesfrom"])
-      assert operationTime.endDate.present
-      assert operationTime.endDate.get() == ZonedDateTime.parse(parameter["operatesuntil"])
-      assert operator == operatorInput
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      operationTime.startDate.present
+      operationTime.startDate.get() == ZonedDateTime.parse(parameter["operatesfrom"])
+      operationTime.endDate.present
+      operationTime.endDate.get() == ZonedDateTime.parse(parameter["operatesuntil"])
+      operator == operatorInput
+      id == parameter["id"]
     }
   }
 
@@ -187,10 +193,10 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert operationTime == OperationTime.notLimited()
-      assert operator == OperatorInput.NO_OPERATOR_ASSIGNED
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      operationTime == OperationTime.notLimited()
+      operator == OperatorInput.NO_OPERATOR_ASSIGNED
+      id == parameter["id"]
     }
   }
 
@@ -211,12 +217,12 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert operationTime.startDate.present
-      assert operationTime.startDate.get() == ZonedDateTime.parse(parameter["operatesfrom"])
-      assert !operationTime.endDate.present
-      assert operator == OperatorInput.NO_OPERATOR_ASSIGNED
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      operationTime.startDate.present
+      operationTime.startDate.get() == ZonedDateTime.parse(parameter["operatesfrom"])
+      !operationTime.endDate.present
+      operator == OperatorInput.NO_OPERATOR_ASSIGNED
+      id == parameter["id"]
     }
   }
 
@@ -237,12 +243,12 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert !operationTime.startDate.present
-      assert operationTime.endDate.present
-      assert operationTime.endDate.get() == ZonedDateTime.parse(parameter["operatesuntil"])
-      assert operator == OperatorInput.NO_OPERATOR_ASSIGNED
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      !operationTime.startDate.present
+      operationTime.endDate.present
+      operationTime.endDate.get() == ZonedDateTime.parse(parameter["operatesuntil"])
+      operator == OperatorInput.NO_OPERATOR_ASSIGNED
+      id == parameter["id"]
     }
   }
 
@@ -264,23 +270,25 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
     input.success
     input.data.get().getClass() == inputClass
     input.data.get().with {
-      assert uuid == UUID.fromString(parameter["uuid"])
-      assert operationTime.startDate.present
-      assert operationTime.startDate.get() == ZonedDateTime.parse(parameter["operatesfrom"])
-      assert operationTime.endDate.present
-      assert operationTime.endDate.get() == ZonedDateTime.parse(parameter["operatesuntil"])
-      assert operator == OperatorInput.NO_OPERATOR_ASSIGNED
-      assert id == parameter["id"]
+      uuid == UUID.fromString(parameter["uuid"])
+      operationTime.startDate.present
+      operationTime.startDate.get() == ZonedDateTime.parse(parameter["operatesfrom"])
+      operationTime.endDate.present
+      operationTime.endDate.get() == ZonedDateTime.parse(parameter["operatesuntil"])
+      operator == OperatorInput.NO_OPERATOR_ASSIGNED
+      id == parameter["id"]
     }
   }
 
   def "An AssetInputFactory should throw an exception on invalid or incomplete fields"() {
     given:
-    def inputFactory = new TestAssetInputFactory()
-    def actualFields = TestAssetInputFactory.newSet("uuid", "operates_from", "operates_until")
+    def actualFields = newSet("uuid", "operates_from", "operates_until")
+
+    // registering fields for the asset
+    ModelFields.register(TestAssetInput, newSet(FieldNamingStrategy.UUID, ID), newSet(OPERATOR, OPERATES_FROM, OPERATES_UNTIL))
 
     when:
-    Try<Void, FactoryException> input = inputFactory.validate(actualFields, TestAssetInput)
+    Try<Void, FactoryException> input = DataSource.validate(actualFields, TestAssetInput)
 
     then:
     input.failure
@@ -311,11 +319,6 @@ class AssetInputEntityFactoryTest extends Specification implements FactoryTestHe
   private static class TestAssetInputFactory extends AssetInputEntityFactory<TestAssetInput, AssetInputEntityData> {
     TestAssetInputFactory() {
       super(TestAssetInput)
-    }
-
-    @Override
-    protected String[] getAdditionalFields() {
-      return new String[0]
     }
 
     @Override